requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
httpx
numpy
scikit-learn